requirements.txt
asgiref==3.4.1
click==8.0.1
cmake==3.21.1.post1
Django==3.2.6
djangorestframework==3.12.4
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.21.1
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
